Refactor: 중첩 router 사용을 위해 layout 구조 변경
