+++ src/layout/MainLayout/index.jsx
@@ -1,14 +1,15 @@
-import GlobalNav from "./GlobalNav"
 
-const HomeLayout = ({children}) => {
+
+
+const MainLayout = ({children}) => {
+
   return (
-    <div className="flex">
-      <GlobalNav />
-      <div className="flex w-full justify-center ml-[72px] mt-4 md:ml-[335px]">
-        {children}
-      </div>
+    // <div className="flex w-full justify-center mt-4 ml-[72px] md:ml-[335px]">
+    <div className=" ml-[72px] md:ml-[335px]">
+      {children}
     </div>
   )
 }
 
-export default HomeLayout
+
+export default MainLayout
