+++ src/pages/Home/index.jsx
@@ -1,29 +1,18 @@
-import Article from "../../components/Article";
-import jsonData from '../../assets/data.json';
-import RecommendUser from "./RecommendUser";
-import Story from "./Story";
-import HomeLayout from "../../layout/MainLayout";
+
+import { Outlet } from "react-router";
+import MainLayout from "../../layout/MainLayout";
+import GlobalNav from "../../layout/GlobalNav";
 
 
 const HomePage = () => {
-
-
-  console.log(jsonData)
   return (
-    <div>
-      <HomeLayout>
-        <main className="w-[630px]">
-          <div className="mb-6">
-            <Story />
-          </div>
-          <Article data={jsonData} />
-          <Article data={jsonData} />
-          <Article data={jsonData} />
-        </main>
-        <div className="hidden md:block pl-20">
-          <RecommendUser />
-        </div>
-      </HomeLayout>
+    <div className="flex">
+      <GlobalNav />
+      <div className="w-full">
+        <MainLayout>
+          <Outlet />
+        </MainLayout>
+      </div>
     </div>
   );
 };
