+++ src/pages/Profile/index.jsx
@@ -1,0 +1,15 @@
+import MainLayout from "../../layout/MainLayout"
+
+const ProfilePage = () => {
+
+  return(
+    <div>
+      <p>asdfs</p>
+
+    </div>
+      // <MainLayout>
+      // </MainLayout>
+  )
+}
+
+export default ProfilePage
